# server/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aegis API
  description: |
    AI-powered security scanning platform — multi-tenant SaaS with per-org data isolation.

    ## Authentication
    Most endpoints require authentication via an `aegis_token` HttpOnly cookie,
    set by the login or register endpoints.

    ## Multi-Tenancy
    Org-scoped endpoints require requests to be made on an org subdomain
    (e.g., `acme.aegis.io`) or a custom domain. The server resolves the org
    from the `Host` header and verifies the user's membership before
    processing the request.
  version: 0.1.0
  contact:
    name: Pixelvide
    url: https://github.com/pixelvide/aegis
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      description: Creates a user, a default org, and sets the auth cookie. Gated by the `signup` feature flag.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration successful
          headers:
            Set-Cookie:
              description: "aegis_token HttpOnly cookie"
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  message:
                    type: string
                    example: "registration successful"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Registration disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      description: Authenticates with email/password. Sets `aegis_token` HttpOnly cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: "aegis_token HttpOnly cookie"
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user
      description: Returns the authenticated user and their organizations.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  orgs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Organizations ─────────────────────────────────────────────────────────
  /orgs:
    post:
      tags: [Organizations]
      summary: Create organization
      description: Creates a new org. The creator becomes the owner.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrgRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [Organizations]
      summary: List organizations
      description: Returns the current user's organizations.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
  /orgs/{slug}:
    get:
      tags: [Organizations]
      summary: Get organization
      description: Returns an organization by slug.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/OrgSlugPath"
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Scans ─────────────────────────────────────────────────────────────────
...
  /dashboard/stats:
    get:
      tags: [Dashboard]
      summary: Dashboard statistics
      description: Returns aggregate statistics for the current org.
      security:
        - cookieAuth: []
      parameters:
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  # ── Projects ──────────────────────────────────────────────────────────────
  /projects:
    post:
      tags: [Projects]
      summary: Create project
      description: Creates a project in the current org.
      security:
        - cookieAuth: []
      parameters:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [Projects]
      summary: List projects
      description: Returns all projects in the current org.
      security:
        - cookieAuth: []
      parameters:
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /projects/{slug}:
    get:
      tags: [Projects]
      summary: Get project
      description: Returns a project by slug.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/OrgSlugPath"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Members ───────────────────────────────────────────────────────────────
  /members:
    get:
      tags: [Members]
      summary: List members
      description: Returns all members of the current org with roles.
      security:
        - cookieAuth: []
      parameters:
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
  /members/invite:
    post:
      tags: [Members]
      summary: Invite member
      description: |
        Invite a user by email. Creates a placeholder account if the user doesn't exist.
        Requires owner or admin role.
      security:
        - cookieAuth: []
      parameters:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteRequest"
      responses:
        "201":
          description: Member invited
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /members/{userId}:
    delete:
      tags: [Members]
      summary: Remove member
      description: Remove a member from the org. Requires owner or admin role. Cannot remove yourself.
      security:
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: User ID to remove
      responses:
        "204":
          description: Member removed
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Config ────────────────────────────────────────────────────────────────
  /config/features:
    get:
      tags: [Config]
      summary: List feature flags
      description: Returns all feature flags.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of feature flags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureFlag"

# ─── Components ───────────────────────────────────────────────────────────────
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: aegis_token
      description: JWT HttpOnly cookie set by login/register
  parameters:
...
    OrgSlugPath:
      name: slug
      in: path
      required: true
      schema:
        type: string
      description: URL-friendly slug
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "invalid request body"

    # ── Auth Schemas ──────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          description: "Must contain uppercase, lowercase, and a digit"
          example: "SecureP@ss1"
        name:
          type: string
          example: "John Doe"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "SecureP@ss1"
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatar_url:
          type: string
        created_at:
          type: string
          format: date-time

    # ── Organization Schemas ──────────────────────────────────────────────
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
          enum: [free, pro, enterprise]
        created_at:
          type: string
          format: date-time
    CreateOrgRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Acme Corp"
        slug:
          type: string
          description: "Auto-generated from name if omitted"
          example: "acme-corp"
        plan:
          type: string
          enum: [free, pro, enterprise]
          default: "free"

    # ── Scan Schemas ──────────────────────────────────────────────────────
...
    Summary:
      type: object
      properties:
        total:
          type: integer
        critical:
          type: integer
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
        info:
          type: integer
...
    Severity:
      type: string
      enum: [critical, high, medium, low, info]
    FindingStatus:
      type: string
      enum: [open, confirmed, fixed, false_positive, wontfix]
    Finding:
      type: object
      properties:
        id:
          type: string
          description: "Finding ID (e.g., AEGIS-001)"
        scan_id:
          type: string
        title:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        cwe:
          type: string
        owasp:
          type: string
        file:
          type: string
        line:
          type: integer
        status:
          $ref: "#/components/schemas/FindingStatus"
        description:
          type: string
          description: "Markdown body"
        exploits:
          type: array
          items:
            $ref: "#/components/schemas/Exploit"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Exploit:
      type: object
      properties:
        id:
          type: string
          format: uuid
        finding_id:
          type: string
        filename:
          type: string
          description: "e.g., exploit.sh, exploit.py"
        language:
          type: string
          description: "e.g., bash, python, javascript"
        code:
          type: string
          description: "Source code of the exploit"
        validated:
          type: boolean

    # ── Dashboard Schemas ─────────────────────────────────────────────────
    DashboardStats:
      type: object
      properties:
        total_scans:
          type: integer
        active_scans:
          type: integer
        total_findings:
          type: integer
        severity_breakdown:
          $ref: "#/components/schemas/Summary"
        recent_findings:
          type: array
          items:
            $ref: "#/components/schemas/Finding"

    # ── Project Schemas ───────────────────────────────────────────────────
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Backend API"
        slug:
          type: string
          description: "Auto-generated from name if omitted"
          example: "backend-api"

    # ── Member Schemas ────────────────────────────────────────────────────
    Member:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatar_url:
          type: string
        role:
          $ref: "#/components/schemas/OrgRole"
    OrgRole:
      type: string
      enum: [owner, admin, member, viewer]
    InviteRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: "new@example.com"
        role:
          $ref: "#/components/schemas/OrgRole"

    # ── Config Schemas ────────────────────────────────────────────────────
    FeatureFlag:
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
        description:
          type: string